example/requirements.txt
alembic==1.2.1
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
jsonpickle==1.2
kafka-python==1.4.7
Mako==1.1.0
MarkupSafe==1.1.1
psycopg2==2.8.3
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.10
task-manager==0.0.1a0
Werkzeug==0.16.0
WTForms==2.2.1
